[api] [tests], make all API CSRF exempt by default if csrf exists
